+++ CMakeLists.txt
@@ -1,12 +1,10 @@
 cmake_minimum_required(VERSION 3.5)
 project(motion)
 
-# Default to C99
 if(NOT CMAKE_C_STANDARD)
   set(CMAKE_C_STANDARD 99)
 endif()
 
-# Default to C++14
 if(NOT CMAKE_CXX_STANDARD)
   set(CMAKE_CXX_STANDARD 14)
 endif()
@@ -15,36 +13,14 @@ if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   add_compile_options(-Wall -Wextra -Wpedantic)
 endif()
 
-# find dependencies
 find_package(ament_cmake REQUIRED)
-find_package(rclcpp REQUIRED)
 find_package(dynamixel_sdk REQUIRED)
 
-add_library(${PROJECT_NAME}
-  "src/${PROJECT_NAME}/serial.cpp")
-
-target_include_directories(${PROJECT_NAME} PUBLIC
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-  $<INSTALL_INTERFACE:include>)
-
-ament_target_dependencies(${PROJECT_NAME} rclcpp dynamixel_sdk)
-
-install(DIRECTORY "include" DESTINATION ".")
-
-install(TARGETS ${PROJECT_NAME}
-  EXPORT export_${PROJECT_NAME}
-  ARCHIVE DESTINATION "lib"
-  LIBRARY DESTINATION "lib"
-  RUNTIME DESTINATION "bin")
-
-add_executable(broadcast_ping "src/broadcast_ping_main.cpp")
-target_include_directories(broadcast_ping PUBLIC
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-  $<INSTALL_INTERFACE:include>)
-target_link_libraries(broadcast_ping ${PROJECT_NAME})
+add_executable(check_joints "src/check_joints_main.cpp")
+ament_target_dependencies(check_joints dynamixel_sdk)
 
 install(TARGETS
-  broadcast_ping
+  check_joints
   DESTINATION lib/${PROJECT_NAME})
 
 if(BUILD_TESTING)
@@ -52,6 +28,5 @@ if(BUILD_TESTING)
   ament_lint_auto_find_test_dependencies()
 endif()
 
-ament_export_include_directories("include")
 ament_export_libraries(${PROJECT_NAME})
 ament_package()
