+++ package.xml
@@ -8,8 +8,7 @@
   <license>MIT License</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>rclcpp</depend>
+  
   <depend>dynamixel_sdk</depend>
 
   <test_depend>ament_lint_auto</test_depend>
